auth observable object, unit and UI tests

+++ nplusone/ContentView.swift
@@ -10,6 +10,9 @@ import CoreData
 
 struct ContentView: View {
     @Environment(\.managedObjectContext) private var viewContext
+    
+    @State private var uiImage: UIImage? = nil
+    @EnvironmentObject var auth: Auth
 
     @FetchRequest(
         sortDescriptors: [NSSortDescriptor(keyPath: \Item.timestamp, ascending: true)],
@@ -17,35 +20,54 @@ struct ContentView: View {
     private var items: FetchedResults<Item>
 
     var body: some View {
-        NavigationView {
-            List {
-                ForEach(items) { item in
-                    NavigationLink {
-                        Text("Item at \(item.timestamp!, formatter: itemFormatter)")
-                    } label: {
-                        Text(item.timestamp!, formatter: itemFormatter)
+        if auth.isLoading {
+            ProgressView()
+                .scaleEffect(2)
+                .frame(maxWidth: .infinity, maxHeight: .infinity)
+                .edgesIgnoringSafeArea(.all)
+        } else {
+            NavigationView {
+                List {
+                    ForEach(items) { item in
+                        NavigationLink {
+                            Text("Item at \(item.timestamp!, formatter: itemFormatter)")
+                        } label: {
+                            Text(item.timestamp!, formatter: itemFormatter)
+                        }
+                    }
+                    .onDelete(perform: deleteItems)
+                }
+                .navigationTitle("N plus ONE")
+                .navigationBarTitleDisplayMode(.inline)
+                .navigationBarItems(
+                    leading: profileImage
+                )
+                .toolbar {
+                    
+                    ToolbarItemGroup(placement: .secondaryAction) {
+                        
+                        EditButton()
+                        Button("Logout") {
+                            logOut()
+                        }
+                    }
+                    ToolbarItem {
+                        Button(action: addItem) {
+                            Label("Add Item", systemImage: "plus")
+                        }
                     }
                 }
-                .onDelete(perform: deleteItems)
-            }
-            .toolbar {
-                ToolbarItem(placement: .navigationBarTrailing) {
-                    EditButton()
-                }
-                ToolbarItem {
-                    Button(action: addItem) {
-                        Label("Add Item", systemImage: "plus")
-                    }
-                }
-            }
-            Text("Select an item")
+                Text("Select an item")
+            }.accessibilityIdentifier("ContentView")
         }
+        
     }
 
     private func addItem() {
         withAnimation {
             let newItem = Item(context: viewContext)
             newItem.timestamp = Date()
+            newItem.user = auth.currentUser
 
             do {
                 try viewContext.save()
@@ -72,6 +94,30 @@ struct ContentView: View {
             }
         }
     }
+    
+    private func logOut() {
+        auth.logout{ success in
+            auth.loggedIn = !success
+        }
+    }
+    
+    var profileImage: some View {
+        Group {
+            if let image = uiImage {
+                Image(uiImage: image)
+                    .resizable()
+                    .aspectRatio(contentMode: .fill)
+                    .frame(width: 36, height: 36)
+                    .clipShape(Circle())
+            } else {
+                Image(systemName: "person.crop.circle")
+                    .resizable()
+                    .aspectRatio(contentMode: .fill)
+                    .frame(width: 36, height: 36)
+                    .clipShape(Circle())
+            }
+        }
+    }
 }
 
 private let itemFormatter: DateFormatter = {
@@ -81,6 +127,6 @@ private let itemFormatter: DateFormatter = {
     return formatter
 }()
 
-#Preview {
-    ContentView().environment(\.managedObjectContext, PersistenceController.preview.container.viewContext)
-}
+//#Preview {
+//    ContentView().environment(\.managedObjectContext, PersistenceController.preview.container.viewContext)
+//}
